+++ scrooge-generator-tests/src/test/resources/gold_file_output_scala/com/twitter/scrooge/test/gold/thriftscala/GoldService.scala
@@ -30,6 +30,7 @@ import org.apache.thrift.TApplicationException
 import scala.collection.mutable.Builder
 import scala.collection.immutable.{Map => immutable$Map, Set => immutable$Set}
 import scala.language.higherKinds
+import scala.reflect.{ClassTag, classTag}
 
 
 @javax.annotation.Generated(value = Array("com.twitter.scrooge.Compiler"))
@@ -284,7 +285,7 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
 
   object DoGreatThings extends ThriftMethod {
     
-    object Args extends ValidatingThriftStructCodec3[Args] {
+    object Args extends ValidatingThriftStructCodec3[Args] with StructBuilderFactory[Args] {
       val NoPassthroughFields: immutable$Map[Short, TFieldBlob] = immutable$Map.empty[Short, TFieldBlob]
       val Struct: TStruct = new TStruct("doGreatThings_args")
       val RequestField: TField = new TField("request", TType.STRUCT, 1)
@@ -309,6 +310,10 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
     
       lazy val structAnnotations: immutable$Map[String, String] =
         immutable$Map.empty[String, String]
+    
+      private val fieldTypes: IndexedSeq[ClassTag[_]] = IndexedSeq(
+        classTag[com.twitter.scrooge.test.gold.thriftscala.Request].asInstanceOf[ClassTag[_]]
+      )
     
       /**
        * Checks that all required fields are non-null.
@@ -335,6 +340,8 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
               com.twitter.scrooge.test.gold.thriftscala.Request.withoutPassthroughFields(field)
             }
         )
+    
+      def newBuilder(): StructBuilder[Args] = new ArgsStructBuilder(_root_.scala.None, fieldTypes)
     
       override def encode(_item: Args, _oproto: TProtocol): Unit = {
         _item.write(_oproto)
@@ -481,11 +488,31 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
       override def productPrefix: String = "Args"
     
       def _codec: ValidatingThriftStructCodec3[Args] = Args
+    
+      def newBuilder(): StructBuilder[Args] = new ArgsStructBuilder(_root_.scala.Some(this), fieldTypes)
+    }
+    
+    private[thriftscala] class ArgsStructBuilder(instance: _root_.scala.Option[Args], fieldTypes: IndexedSeq[ClassTag[_]])
+        extends StructBuilder[Args](fieldTypes) {
+    
+      def build(): Args = instance match {
+        case _root_.scala.Some(i) =>
+          Args(
+            (if (fieldArray(0) == null) i.request else fieldArray(0)).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.Request]
+          )
+        case _root_.scala.None =>
+          if (fieldArray.contains(null)) throw new InvalidFieldsException(structBuildError("Args"))
+          else {
+            Args(
+              fieldArray(0).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.Request]
+            )
+          }
+        }
     }
 
     type SuccessType = com.twitter.scrooge.test.gold.thriftscala.Response
     
-    object Result extends ValidatingThriftStructCodec3[Result] {
+    object Result extends ValidatingThriftStructCodec3[Result] with StructBuilderFactory[Result] {
       val NoPassthroughFields: immutable$Map[Short, TFieldBlob] = immutable$Map.empty[Short, TFieldBlob]
       val Struct: TStruct = new TStruct("doGreatThings_result")
       val SuccessField: TField = new TField("success", TType.STRUCT, 0)
@@ -524,6 +551,11 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
       lazy val structAnnotations: immutable$Map[String, String] =
         immutable$Map.empty[String, String]
     
+      private val fieldTypes: IndexedSeq[ClassTag[_]] = IndexedSeq(
+        classTag[_root_.scala.Option[com.twitter.scrooge.test.gold.thriftscala.Response]].asInstanceOf[ClassTag[_]],
+        classTag[_root_.scala.Option[com.twitter.scrooge.test.gold.thriftscala.OverCapacityException]].asInstanceOf[ClassTag[_]]
+      )
+    
       /**
        * Checks that all required fields are non-null.
        */
@@ -559,6 +591,8 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
               }
             }
         )
+    
+      def newBuilder(): StructBuilder[Result] = new ResultStructBuilder(_root_.scala.None, fieldTypes)
     
       override def encode(_item: Result, _oproto: TProtocol): Unit = {
         _item.write(_oproto)
@@ -753,6 +787,28 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
       override def productPrefix: String = "Result"
     
       def _codec: ValidatingThriftStructCodec3[Result] = Result
+    
+      def newBuilder(): StructBuilder[Result] = new ResultStructBuilder(_root_.scala.Some(this), fieldTypes)
+    }
+    
+    private[thriftscala] class ResultStructBuilder(instance: _root_.scala.Option[Result], fieldTypes: IndexedSeq[ClassTag[_]])
+        extends StructBuilder[Result](fieldTypes) {
+    
+      def build(): Result = instance match {
+        case _root_.scala.Some(i) =>
+          Result(
+            (if (fieldArray(0) == null) i.success else fieldArray(0)).asInstanceOf[_root_.scala.Option[com.twitter.scrooge.test.gold.thriftscala.Response]],
+            (if (fieldArray(1) == null) i.ex else fieldArray(1)).asInstanceOf[_root_.scala.Option[com.twitter.scrooge.test.gold.thriftscala.OverCapacityException]]
+          )
+        case _root_.scala.None =>
+          if (fieldArray.contains(null)) throw new InvalidFieldsException(structBuildError("Result"))
+          else {
+            Result(
+              fieldArray(0).asInstanceOf[_root_.scala.Option[com.twitter.scrooge.test.gold.thriftscala.Response]],
+              fieldArray(1).asInstanceOf[_root_.scala.Option[com.twitter.scrooge.test.gold.thriftscala.OverCapacityException]]
+            )
+          }
+        }
     }
 
     val annotations: immutable$Map[String, String] = immutable$Map(
@@ -798,7 +854,7 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
 
   object NoExceptionCall extends ThriftMethod {
     
-    object Args extends ValidatingThriftStructCodec3[Args] {
+    object Args extends ValidatingThriftStructCodec3[Args] with StructBuilderFactory[Args] {
       val NoPassthroughFields: immutable$Map[Short, TFieldBlob] = immutable$Map.empty[Short, TFieldBlob]
       val Struct: TStruct = new TStruct("noExceptionCall_args")
       val RequestField: TField = new TField("request", TType.STRUCT, 1)
@@ -823,6 +879,10 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
     
       lazy val structAnnotations: immutable$Map[String, String] =
         immutable$Map.empty[String, String]
+    
+      private val fieldTypes: IndexedSeq[ClassTag[_]] = IndexedSeq(
+        classTag[com.twitter.scrooge.test.gold.thriftscala.Request].asInstanceOf[ClassTag[_]]
+      )
     
       /**
        * Checks that all required fields are non-null.
@@ -849,6 +909,8 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
               com.twitter.scrooge.test.gold.thriftscala.Request.withoutPassthroughFields(field)
             }
         )
+    
+      def newBuilder(): StructBuilder[Args] = new ArgsStructBuilder(_root_.scala.None, fieldTypes)
     
       override def encode(_item: Args, _oproto: TProtocol): Unit = {
         _item.write(_oproto)
@@ -995,11 +1057,31 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
       override def productPrefix: String = "Args"
     
       def _codec: ValidatingThriftStructCodec3[Args] = Args
+    
+      def newBuilder(): StructBuilder[Args] = new ArgsStructBuilder(_root_.scala.Some(this), fieldTypes)
+    }
+    
+    private[thriftscala] class ArgsStructBuilder(instance: _root_.scala.Option[Args], fieldTypes: IndexedSeq[ClassTag[_]])
+        extends StructBuilder[Args](fieldTypes) {
+    
+      def build(): Args = instance match {
+        case _root_.scala.Some(i) =>
+          Args(
+            (if (fieldArray(0) == null) i.request else fieldArray(0)).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.Request]
+          )
+        case _root_.scala.None =>
+          if (fieldArray.contains(null)) throw new InvalidFieldsException(structBuildError("Args"))
+          else {
+            Args(
+              fieldArray(0).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.Request]
+            )
+          }
+        }
     }
 
     type SuccessType = com.twitter.scrooge.test.gold.thriftscala.Response
     
-    object Result extends ValidatingThriftStructCodec3[Result] {
+    object Result extends ValidatingThriftStructCodec3[Result] with StructBuilderFactory[Result] {
       val NoPassthroughFields: immutable$Map[Short, TFieldBlob] = immutable$Map.empty[Short, TFieldBlob]
       val Struct: TStruct = new TStruct("noExceptionCall_result")
       val SuccessField: TField = new TField("success", TType.STRUCT, 0)
@@ -1024,6 +1106,10 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
     
       lazy val structAnnotations: immutable$Map[String, String] =
         immutable$Map.empty[String, String]
+    
+      private val fieldTypes: IndexedSeq[ClassTag[_]] = IndexedSeq(
+        classTag[_root_.scala.Option[com.twitter.scrooge.test.gold.thriftscala.Response]].asInstanceOf[ClassTag[_]]
+      )
     
       /**
        * Checks that all required fields are non-null.
@@ -1052,6 +1138,8 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
               }
             }
         )
+    
+      def newBuilder(): StructBuilder[Result] = new ResultStructBuilder(_root_.scala.None, fieldTypes)
     
       override def encode(_item: Result, _oproto: TProtocol): Unit = {
         _item.write(_oproto)
@@ -1200,6 +1288,26 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
       override def productPrefix: String = "Result"
     
       def _codec: ValidatingThriftStructCodec3[Result] = Result
+    
+      def newBuilder(): StructBuilder[Result] = new ResultStructBuilder(_root_.scala.Some(this), fieldTypes)
+    }
+    
+    private[thriftscala] class ResultStructBuilder(instance: _root_.scala.Option[Result], fieldTypes: IndexedSeq[ClassTag[_]])
+        extends StructBuilder[Result](fieldTypes) {
+    
+      def build(): Result = instance match {
+        case _root_.scala.Some(i) =>
+          Result(
+            (if (fieldArray(0) == null) i.success else fieldArray(0)).asInstanceOf[_root_.scala.Option[com.twitter.scrooge.test.gold.thriftscala.Response]]
+          )
+        case _root_.scala.None =>
+          if (fieldArray.contains(null)) throw new InvalidFieldsException(structBuildError("Result"))
+          else {
+            Result(
+              fieldArray(0).asInstanceOf[_root_.scala.Option[com.twitter.scrooge.test.gold.thriftscala.Response]]
+            )
+          }
+        }
     }
 
     val annotations: immutable$Map[String, String] = immutable$Map.empty